Setting up types
